was/pyproject.toml
[tool.poetry]
name = "avikus-backend-assignment"
version = "0.1.0"
description = "avikus_backend_assignment"
authors = ["ShinHeeJin <sisesu@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
marshmallow = "3.20.1"
gunicorn = "21.2"
python-dotenv = "1.0.0"
pytest = "7.4.2"
pytest-cov = "4.1.0"
flask-caching = "1.10.1"
sqlalchemy = "1.4.40"
mysqlclient = "2.1.1"
flask-sqlalchemy = "2.5.1"
flask-restx = "^1.2.0"
redis = "^5.0.1"
eventlet = "^0.33.3"
mysql-connector-python = "8.0.29"
celery = "5.2.3"
flask = "2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
